requirements.txt
streamlit
scikit-learn
pandas
plotly-express
